[cache] Code refactoring
