+++ cache/memory/examples_test.go
@@ -1,4 +1,4 @@
-package cache
+package memory
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
